+++ packages/components/src/components/banner/banner-styles.ts
@@ -5,22 +5,24 @@ import {
   gridExtendedOffsetBase,
   gridExtendedOffsetS,
   gridExtendedOffsetXXL,
+  motionDurationModerate,
+  motionDurationLong,
+  motionEasingIn,
+  motionEasingOut,
 } from '@porsche-design-system/utilities-v2';
 import { getCss } from '../../utils';
 import { BANNER_Z_INDEX } from '../../constants';
-import { addImportantToEachRule, colorSchemeStyles, hostHiddenStyles } from '../../styles';
+import {
+  addImportantToEachRule,
+  colorSchemeStyles,
+  cssVariableAnimationDuration,
+  hostHiddenStyles,
+} from '../../styles';
 
 const cssVariableTop = '--p-banner-position-top';
 const cssVariableBottom = '--p-banner-position-bottom';
-const cssVariableAnimationDuration = '--p-animation-duration';
 const cssVariableZIndex = '--p-internal-banner-z-index';
 
-export const ANIMATION_DURATION = 600;
-
-const duration = `var(${cssVariableAnimationDuration},${ANIMATION_DURATION}ms)`;
-
-const easeInQuad = 'cubic-bezier(0.45,0,0.55,1)';
-const easeOutQuad = 'cubic-bezier(0.5,1,0.89,1)';
 const topBottomFallback = '56px';
 
 export const getComponentCss = (isOpen: boolean): string => {
@@ -43,14 +45,14 @@ export const getComponentCss = (isOpen: boolean): string => {
               opacity: 1,
               visibility: 'inherit',
               transform: 'translate3d(0,0,0)',
-              transition: `opacity ${duration} ${easeInQuad},transform ${duration} ${easeInQuad}`,
+              transition: `transform var(${cssVariableAnimationDuration}, ${motionDurationModerate}) ${motionEasingIn}, opacity var(${cssVariableAnimationDuration}, ${motionDurationModerate}) ${motionEasingIn}`,
             }
           : {
               opacity: 0,
               visibility: 'hidden',
               transform: `translate3d(0,calc(var(${cssVariableBottom},${topBottomFallback}) + 100%),0)`,
               '&(.hydrated),&(.ssr)': {
-                transition: `visibility 0s linear ${duration},opacity ${duration} ${easeOutQuad},transform ${duration} ${easeOutQuad}`,
+                transition: `visibility 0s linear var(${cssVariableAnimationDuration}, ${motionDurationLong}), transform var(${cssVariableAnimationDuration}, ${motionDurationModerate}) ${motionEasingOut}, opacity var(${cssVariableAnimationDuration}, ${motionDurationModerate}) ${motionEasingOut}`,
               },
             }),
         [getMediaQueryMin('s')]: {